arm64: Log the machine name string during boot up
The change to refactor kernel/setup.c to use the common
of_flat_dt_get_machine_name() API has apparently removed
the line which prints the device tree model string during
boot. Having the model string in the kernel log is helpful,
so add it back in. This change was already merged in past
but possibly got overridden during upmerge. Add it back.
While at it, add back print for the processor name and
its rev id as well.

Change-Id: I7dccc3ab00f5b67753cdd256846a522596c5058f
Signed-off-by: Stepan Moskovchenko <stepanm@codeaurora.org>
Signed-off-by: Kaushal Kumar <kaushalk@codeaurora.org>