irqchip: gic: Fix msm_show_resume_irq for 64-bit
msm_show_resume_irq function uses unsigned long to
capture the 32-bit values from the irq pending register.
This then is used to identify the pending interrupts using
find_next_bit and results in an incorrect value on 64-bit
targets as the 32-bit data is encapusulated in a 64-bit register.

Change-Id: Id8197ebdfa7d1701822892c5f7369ae82cfe35cf
Signed-off-by: Rohit Vaswani <rvaswani@codeaurora.org>
Signed-off-by: Prasad Sodagudi <psodagud@codeaurora.org>