UPSTREAM: efi: efistub: allow allocation alignment larger than EFI_PAGE_SIZE
On systems with 64 KB pages, it is preferable for UEFI memory map
entries to be 64 KB aligned multiples of 64 KB, because it relieves
us of having to deal with the residues.
So, if EFI_ALLOC_ALIGN is #define'd by the platform, use it to round
up all memory allocations made.

Acked-by: Matt Fleming <matt.fleming@intel.com>
Acked-by: Borislav Petkov <bp@suse.de>
Tested-by: Leif Lindholm <leif.lindholm@linaro.org>
Signed-off-by: Ard Biesheuvel <ard.biesheuvel@linaro.org>

(cherry picked from commit cf2b0f102cdf912eedb87b10271fa0ad582cf2c1)
Signed-off-by: Tomasz Figa <tfiga@chromium.org>

Bug: 24475017

Change-Id: I79fedd4b366bddc5f2aac25d27f04ba3963a0612
Signed-off-by: Kees Cook <keescook@google.com>