thermal: tsens: Log threshold states
Upper, lower and critical thresholds are set by clients for each sensor
to monitor the temperature. Clients will be notified when there is
a cross over in the thresholds based on the state. Log upper, lower
and critical threshold value and state to ensure that thresholds set by
clients are reflected properly.

Change-Id: I1f388219dbb839de2e441c85e3fa2c53759a5973
Signed-off-by: Rama Krishna Phani A <rphani@codeaurora.org>