msm: kgsl: Clear internal timestamps after a fault
Sometimes a fault can occur in the various bits that happen after a
user command is retired - if this happens there might be contexts
waiting to detach on internal ringbuffer commands that are not going
to be executed. All internal commands are not replayed on fault
recovery so the memstore needs to be updated and the events need
to be kicked so that the accounting is correct.

Change-Id: Ic0dedbad5de0e8350b60533871e16ffabdcb106b
Signed-off-by: Jordan Crouse <jcrouse@codeaurora.org>