Documentation:dt:crypto:msm: Clarification of qcom,ce-device attribute
qcom,ce-device attribute is defined with a encoding scheme to
support multiple instances of fde and pfe. The encoding scheme provides
compatiblity to the current single instance implementation for all targets.
For the existing targets of single instance of fde and pfe, no need
to change DT file.

Furthermore, add clarification to qcom,ce-opp-freq and qcom,msm_bus,vectors
attributes to match the actual usage. Those two attributes are optional,
not mandatory.  Though, they are required for msm targets,
not for fsm targets.

Change-Id: I853cbe0a26285ad570db6855a28c4cf6159073b1
Acked-by: Che-Min Hsieh <cheminh@qti.qualcomm.com>
Signed-off-by: Zhen Kong <zkong@codeaurora.org>