power: smb1351-charger: detect HVDCP
The HVDCP algorithm runs after a DCP has been detected by APSD.
Currently we ignore the results of the HVDCP algorithm and
report HVDCP chargers as DCP to the USB power supply.

Read the results of the HVDCP algorithm and notify the USB power
supply if an HVDCP charger has been detected.

CRs-Fixed: 911532
Change-Id: I70415bd90f9616fcf94a1a8187141082e7ae47c1
Signed-off-by: Nicholas Troast <ntroast@codeaurora.org>