usb: dwc3: msm: Allow VBUS regulator to be optional
Host-only mode can be supported by directly connecting the VBUS
pin on the USB connector to an always-on 5V supply. In this case
no explicit control from SW is needed. Hence treat the regulator
as optional: if regulator_get fails, allow host operation to
successfully continue assuming that VBUS is supplied externally.

Change-Id: I1b14d3542bd9b00f6631ddea6471b588193de8dc
Signed-off-by: Jack Pham <jackp@codeaurora.org>

Conflicts:
	drivers/usb/dwc3/dwc3-msm.c