clk: msm: Fix a bug in the derive_device_list api
If the cpu device_node is found and added to the device_list,
do not attempt to find the platform_device node. Also, initialize
the device_list entries to null at the beginning of the for loop
to prevent the if check from evaluating to true even when
get_cpu_device fails or the dev_node does not match the cpu_node.

Change-Id: If099f78c6582ca67e9e00dd271d0b52dd67c2157
Signed-off-by: Devesh Jhunjhunwala <deveshj@codeaurora.org>