ARM: dts: msm: switch to RPM control for regulators on msmtitanium
Convert most of the msmtitanium stub-regulator devices to a
rpm-smd-regulator devices. This ensures that requests made for
these regulators are aggregated by the RPM processor along
with the requests from other processors.

SMPS5/6 should not be converted to an rpm-smd-regulator device
because it is used and controlled by the CPR regulator device.
Also, add a dummy apc and gfx corner regulator.

Change-Id: Id4d9be8b18a27193bcb2c70c0b6eb377f609164f
Signed-off-by: Ashay Jaiswal <ashayj@codeaurora.org>