msm: mhi_dev: Update MHI interrupt and local address properties
Modem Host Interface (MHI) is not required to register for a
dedicated interrupt for control and data processing since the
interrupt is handled by PCIe driver. MHI registers with the
PCIe driver and gets notification to handle control and data
processing with the host. MHI uses IPA DMA for transferring
data between host and device and does not need the local
physical address property which was used to map the host
address region and address translation. Therefore remove
these properties.

Change-Id: Ia108347b4a9d5d031676686b2de5f9dac21602a5
Signed-off-by: Siddartha Mohanadoss <smohanad@codeaurora.org>