ASoC: wcd: set maximum limit for impedance for lineout detection
Due to slow insertion of headset, it can get detected as headphone
momentarily and the impedance is also very high which makes the
headset to be detected as lineout. Set a maximum limit for impedance
and do not report it as lineout in report plug if the value is very
high.

Change-Id: I3257c8f117fdb596a4aa9e743d39d517ad109c79
Signed-off-by: Simmi Pateriya <simmip@codeaurora.org>

Conflicts:
	sound/soc/codecs/wcd-mbhc-v2.c
	sound/soc/codecs/wcd-mbhc-v2.h