power: qpnp-charger: Use VCHG to read input current
The VCHG can be configured to track the current drawn by the
charger at its input (IUSB). The voltage at VCHG is proportional
to the IUSB.

PMI8950 - VCHG is internally routed to a VADC channel.
PMI8994 - The VCHG pin can be externally routed to VADC via a MPP

CRs-Fixed: 930343
Change-Id: I4c299c92d6405f65f1280655c8a5d004ef401d38
Signed-off-by: Anirudh Ghayal <aghayal@codeaurora.org>

Conflicts:
	drivers/power/qpnp-smbcharger.c