usb: f_gsi: Block doorbell to GSI when USB is disconnected
It is possible that USB h/w rings the doorbell to GSI
after IPA clocks are turned OFF during disconnect handling.
Add a fix to block the doorbell on cable disconnect and
re-enable the doorbell when data path is enabled to avoid
this potential race condition between h/w and s/w.

Change-Id: I60fd753787605d751bbb3bba355a89e96b7b1eb3
Signed-off-by: Devdutt Patnaik <dpatnaik@codeaurora.org>