usb: dwc3-msm: Make power collapse and power-on-reset mandatory
Upon cable disconnect, the controller is allowed to enter the
lowest possible power state, which may involve power collapse
via GDSC. Even if GDSC is unavailable on some configurations,
when powering back up the controller it is wise to perform the
full power-on-reset sequence.

Currently these behaviors are optional and enabled via device
tree properties. There is no reason not to perform this sequence
always so make them mandatory and remove the DT properties.

Change-Id: I9b233402c90383782c53732d54e76360f4ea1b55
Signed-off-by: Jack Pham <jackp@codeaurora.org>

Conflicts:
	drivers/usb/dwc3/dwc3-msm.c