usb: dwc3-msm: dbm: Change API to direct calls
Now that there is only a single DBM driver, get rid of the
function pointer based APIs and simply expose the functions
to be called directly by dwc3-msm. From this we can consolidate
the dbm and dbm_data structures and hide the definition in dbm.c.

Also remove kfree() calls since devm_kzalloc() is being used. The
memory will be automatically freed when the device is unbound.

Change-Id: Ia9080ccaf36766ce58655364b247aa0025de7670
Signed-off-by: Jack Pham <jackp@codeaurora.org>