usb: dwc3-msm: Call pm_runtime_get/put on the parent device
Currently in start_peripheral() / start_host() the pm_runtime
get_sync() and put_sync() operations are being performed on the
dwc device. Since it is the mdwc parent device which actually
implements the callbacks to manage clocks and other resources,
it is more correct to ensure that the parent is not in suspend
to avoid any unclocked accesses.

Also replace the awkward one second delay in the A_HOST state
with proper PM autosuspend. This requires mark_last_busy() to be
called in start_host().

Change-Id: Ia8ae158f9ec7e4c91546f889892560df4b722735
Signed-off-by: Jack Pham <jackp@codeaurora.org>

Conflicts:
	drivers/usb/dwc3/dwc3-msm.c