usb: dwc3-msm: Enable USB3 PHY suspend in host mode
When starting host mode ensure that USB3 PHY is allowed to
suspend via GUSB3PIPECTL SUSPHY bit. This will allow the
PHY to enter P3 state, and is a precondition for allowing
controller suspend when a superspeed peripheral is attached
and itself is allowed to suspend (such as a SS hub). In
addition this will allow for additional power savings in HS
mode as the SS PHY will always be in low power mode.

While at it remove the unused 'hcd' local variable.

Change-Id: I4545aa959b7d78b88011f6afb80da11934b85a67
Signed-off-by: Jack Pham <jackp@codeaurora.org>