msm:ipa3: fix dependencies not added on composition switch
USB\Q6 dependencies are not added on composition switch 9063->90A9.
This is due to tethering bridge state machine getting out of sync when
both RMNET and MBIM are in the same USB composition, causing the
dependencies not to be added in the tethering bridge INIT phase.
This change removes the tethering bridge state machine and moves
the addition of dependencies to the CONNECT phase.

Change-Id: If44948b201872dad3319c203a4b278fc81a101b0
Signed-off-by: Amir Levy <alevy@codeaurora.org>