usb: gadget: f_gsi: Add logging support
Use the generic logging framework to log the messages from
f_gsi driver to console and also to a debug buffer. This will
help in getting the debug logs when the issue is seen,
instead of having to enable dynamic_debug and then reproduce
the issue again.
The logs can be read from:

     cat /sys/kernel/debug/debug/read_buf

Change-Id: I4809fd4f8113abf7bfd63211204478d0dc90fe12
Signed-off-by: Azhar Shaikh <azhars@codeaurora.org>