soc: peripheral-loader: only reclaim memory if it was assigned
In the error path of pil_boot, memory is attempted to be reclaimed
from the source VMID of both Linux and the subsystem. This may lead
to a panic if this attempted during one of the error conditions
previous to the assign call of the region for both Linux and the
subsystem.

Fix this by introducing a flag to track the state of the hypervisor
assignment.

Change-Id: I91ab134e645718c1c7c19969da80b879b53af6fa
CRs-Fixed: 957289
Signed-off-by: David Keitel <dkeitel@codeaurora.org>

Conflicts:
	drivers/soc/qcom/peripheral-loader.c