power: qpnp-smbcharger: configure the start of ICL
An overshoot in the USB input current is observed during charger
buck startup. The ICL control is not active when the buck starts
and if there is high load the buck pulls as much current as it
needs.

The root cause is that the ICL control is enabled only after soft
start is done, that is when APSD completes.

The fix is to enable ICL as soon as the buck starts switching.
This can be done by configuring ICL to start based on Syson LDO.
Syson LDO is used to power the drivers of the switching FETS and
becomes active right when the buck starts switching.

Also, during OTG, the ICL control buffer requires voltage sourced
from USB mid. On charger insertion, USB mid has 5V before
switching begins however for OTG, USB mid doesn't have 5V until
the boost starts. If ICL control is left at Syson LDO, it will
falsely detect a short circuit and disables boost. The boost will
never start again when this happens. So reset the ICL control to
start upon soft start completion before starting boost mode.

CRs-Fixed: 962458
Change-Id: I48869f3fa04d0f415bd5a7eb201bd86c7339a362
Signed-off-by: Abhijeet Dharmapurikar <adharmap@codeaurora.org>
Signed-off-by: Subbaraman Narayanamurthy <subbaram@codeaurora.org>

Conflicts:
	drivers/power/qpnp-smbcharger.c