ARM: dts: msm: fix low FPS issue on Sharp FHD panel
On 8996 hardware, every alternate frame
on the Sharp FHD panel is skipped since the
vsync is exactly alligned to the frame bounder and
their is no additional buffer time that is needed
for configuration and kickoff. TO fix this,
run the DSI pixel clock at a higher transfer rate
to reduce the transfer time.

Change-Id: I8443f450b6b0c0b2b66592bbb55b167221d2f342
Signed-off-by: Chandan Uddaraju <chandanu@codeaurora.org>