ARM: dts: msm: Push C4 with PSCI to 11.5ms for MSM8996
C4 with psci overhead is four times that of C4 without psci state.
Based on power profile across system wide usecase, C4 with psci
proves to be costly as the number of pre-mature wakeups from
fpc is huge.
This change pushes C4 with psci state to 11.5ms breakeven point. The
idea is to exercise C4 without psci till that point and allow C4 with
psci only for D4 case.

Change-Id: I81902ac2ff0a0e99efc21d169be472952dcfd3e2
Signed-off-by: Mahesh Sivasubramanian <msivasub@codeaurora.org>
Signed-off-by: Girish S Ghongdemath <girishsg@codeaurora.org>