ARM: dts: msm: Add BAM DMUX no-cpu-affinity for MSM8937
BAM DMUX workqueue is bound to CPU0 to support the network stack
requirements to avoid the out of order packet issue.

In MSM8937 CPU0 is power CPU and bounding workqueue to CPU0 is causing
more power leak in system. Remove the CPU affinity to let the scheduler
to decide the CPU for BAM DMUX work queue to save the power and enable
the RPS feature in network stack to support out of order packet issue.

CRs-Fixed: 970191
Change-Id: I4dd1450f1cf669d63877dd57a047fc939c48bca6
Signed-off-by: Arun Kumar Neelakantam <aneela@codeaurora.org>