msm: vidc: Update instance state before cap check
Currently when capabilities check is done, instance
is not yes moved to OPEN_DONE state. This means that
the capabilties are checked against defalut values
instead of the values returned by Venus. Hence move
the instance to OPEN_DONE state to make sure that the
capabilities are filled properly in session_init_done.

CRs-Fixed: 950899
Change-Id: I5868980e19bd2e1b9e3dbb724b010c0c6ce5e845
Signed-off-by: Praneeth Paladugu <ppaladug@codeaurora.org>