usb: gadget: f_gsi: Remove gadget restart while composition switch
Commit b474e6d31c5d3843 ("Configure fixed GSI endpoints and related
EP OPs") allocated fixed endpoints for functions that use GSI as a
transport.The call to gadget restart was done in bind_config, since
a controller restart was recommended when we enable/disable GSI on
different endpoints. Now with fixed GSI endpoints a controller
restart is not needed anymore.

CRs-Fixed: 973625
Change-Id: Ia440794d37a91ec7b52fb3dbf3819df4e22b60a5
Signed-off-by: Azhar Shaikh <azhars@codeaurora.org>