power: smb1351-charger: Add control for parallel charger configuration
SMB1351 can be configured as the main charger in a parallel charging
solution. In the main chargers role, it is responsible for detecting the
charger adapter's input current limit(ICL) and splitting the ICL with the
slave charger. It also splits the full charging current(FCC) and helps
reduce the thermal dissipation across both the chargers.

CRs-Fixed: 991446
Change-Id: If6040cc9b49019736f267da628d95f1078a66c6f
Signed-off-by: Fenglin Wu <fenglinw@codeaurora.org>

Conflicts:
	drivers/power/smb1351-charger.c