USB: gadget: Unconfigure bam2bam endpoints during suspend
If remotewakeup is not enabled or allowed, USB driver disconnects
BAM2BAM connection with IPA as part of suspend and reconnect BAM2BAM
pipes during resume. As part of suspend, it is not calling
msm_ep_unconfig() before performing BAM2BAM pipes disconnection. Due to
this, after resume, configuring DBM endpoint fails and leads to USB
transfers not working. Hence fix this by unconfiguring DBM endpoint
in suspend if remotewakeup is not supported or enabled.

Change-Id: Ifc68fc91f46ca34c97c7808c4ebec152bc2e7e86
Signed-off-by: Vijayavardhan Vennapusa <vvreddy@codeaurora.org>