msm: mdss: delay dma commands for split-dsi cmd mode panels
For split-dsi command mode panels, when burst mode
is enabled, driver needs to make sure that dsi commands
are not sent at the beginning of the frame in order
to avoid dsi fifo underflows.
Prevent this condition by making sure that commands
are delayed after the rd pointer interrupt is received.

Change-Id: I5f78efc88dc08da06e24c101ad1e8ebe693db7c7
Signed-off-by: Ingrid Gallardo <ingridg@codeaurora.org>

Conflicts:
	drivers/video/msm/mdss/mdss_mdp_intf_cmd.c