Revert "soc: qcom: ssr: Force shutdown of modem during adb reboot"
This reverts 'commit 2d83dfa6925d ("soc: qcom: ssr: Force shutdown
of modem during adb reboot")' which was raised as result of customer
requirement. It made way into mainline with a expectation that similar
issues on other platform can be tackled with this change or may be
required by subsystems' for quick housekeeping before SoC reset.

Here we set the restart bit but do not wait for ACK from other subsystem.
Due to this, we initiate our SoC shutdown before the other subsystem is
ready to shutdown which is cause of present issue so reverting this now.

Change-Id: Iadec546ba227492756b304717358a5dc7b53ef18
Signed-off-by: Avaneesh Kumar Dwivedi <akdwived@codeaurora.org>