power: qpnp-fg: Clear VACT_INT_ERROR after IMA error recovery
Currently when the IMA error recovery sequence is executed, we
restore some of the SRAM registers backed up. This includes
VACT_INT_ERROR as well. However, we have to clear VACT_INT_ERROR
for a good starting point after resetting FG. This will help in
accuracy of the SOC reported.

CRs-Fixed: 1039053
Change-Id: I2f908481ae2a265efb3fa29193f5a87fea096449
Signed-off-by: Subbaraman Narayanamurthy <subbaram@codeaurora.org>