usb: phy: qusb: Fix typo with emulation related property name
Driver is using "qcom,emu-init-seq" for finding length of property
but using "qcom,qemu-init-seq" instead of "qcom,emu-init-seq" to
get passed values from device tree. With this, length of property
comes as expected but values are not received. This results into
programming zeros as values with offset as zero instead of expected
<value, offset> passed through devicetree.

CRs-Fixed: 1051115
Change-Id: I1809fcef844d275175814e636591b87e91432609
Signed-off-by: Mayank Rana <mrana@codeaurora.org>